// src/interfaces/consultantTemplate.ts
export interface createSourceTemplete {
    project_title: string,
    thematic_area: string,
    years_of_experience: string
    consultant_designation: string
    project_details: string
}


export interface teamSourceProjectTemplate {
    project_title: string,
    thematic_area: string,
    client_budget: string
    start_date: string
    project_duration: string
    consultant_project_location: string,
    project_details: string
}


export interface contractConsultantTemplate {
    project_title: string,
    thematic_area: string,
    years_of_experience: string
    consultant_designation: string
    project_location: string
    project_details: string,
    consultant_duration: string
    commitment_level: string
    work_type: string
    start_date: string
    client_budget: string
}